# tmp/swagger.json
{
	"openapi": "3.0.0",
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"SendTelegramMessageRequest": {
				"description": "Интерфейс запроса для отправки сообщения в Telegram.\nОпределите необходимые поля, например:\n- chatId: string;\n- text: string;",
				"properties": {
					"chatId": {
						"type": "string"
					},
					"text": {
						"type": "string"
					}
				},
				"required": [
					"chatId",
					"text"
				],
				"type": "object",
				"additionalProperties": true
			},
			"UserDTO": {
				"properties": {
					"telegramId": {
						"type": "string"
					},
					"firstName": {
						"type": "string"
					},
					"lastName": {
						"type": "string"
					},
					"username": {
						"type": "string"
					},
					"photoUrl": {
						"type": "string"
					},
					"coins": {
						"type": "number",
						"format": "double"
					},
					"role": {
						"type": "string",
						"enum": [
							"user",
							"admin",
							"moderator"
						]
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"telegramId",
					"createdAt"
				],
				"type": "object",
				"additionalProperties": true
			},
			"MovieDTO": {
				"properties": {
					"title": {
						"type": "string"
					},
					"author": {
						"$ref": "#/components/schemas/UserDTO"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"isDeleted": {
						"type": "boolean"
					},
					"deletedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"title",
					"author",
					"createdAt",
					"isDeleted"
				],
				"type": "object",
				"additionalProperties": true
			},
			"FinalResultsRequest": {
				"properties": {
					"firstPlace": {
						"type": "string"
					},
					"secondPlace": {
						"type": "string"
					},
					"thirdPlace": {
						"type": "string"
					}
				},
				"required": [
					"firstPlace"
				],
				"type": "object",
				"additionalProperties": true
			},
			"TelegramAuthRequest": {
				"description": "Интерфейс запроса для аутентификации через Telegram Web App",
				"properties": {
					"initData": {
						"type": "string"
					}
				},
				"required": [
					"initData"
				],
				"type": "object",
				"additionalProperties": true
			},
			"CreateAffirmationRequest": {
				"properties": {
					"text": {
						"type": "string"
					}
				},
				"required": [
					"text"
				],
				"type": "object",
				"additionalProperties": true
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "hohma-server",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
		"/telegram/send": {
			"post": {
				"operationId": "SendTelegramMessageHandler",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Отправка сообщения в Telegram",
				"tags": [
					"Telegram"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"description": "Данные сообщения",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/SendTelegramMessageRequest",
								"description": "Данные сообщения"
							}
						}
					}
				}
			}
		},
		"/movies": {
			"get": {
				"operationId": "GetMovies",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/MovieDTO"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"description": "Получить список всех фильмов",
				"tags": [
					"Movies"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "all",
						"required": false,
						"schema": {
							"type": "boolean"
						}
					}
				]
			},
			"post": {
				"operationId": "CreateMovie",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MovieDTO"
								}
							}
						}
					}
				},
				"description": "Создать новый фильм",
				"tags": [
					"Movies"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {}
						}
					}
				}
			}
		},
		"/movies/{id}": {
			"get": {
				"operationId": "GetMovieById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/MovieDTO"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"description": "Получить фильм по ID",
				"tags": [
					"Movies"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateMovie",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/MovieDTO"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"description": "Обновить информацию о фильме",
				"tags": [
					"Movies"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {}
						}
					}
				}
			},
			"delete": {
				"operationId": "SoftDeleteMovie",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Мягкое удаление фильма",
				"tags": [
					"Movies"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/movies/{id}/permanent": {
			"delete": {
				"operationId": "DeleteMoviePermanently",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Полное удаление фильма",
				"tags": [
					"Movies"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/finalResults": {
			"post": {
				"operationId": "AddFinalResults",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"description": "Добавить новые результаты",
				"tags": [
					"FinalResults"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"description": "Объект с данными для добавления результатов",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/FinalResultsRequest",
								"description": "Объект с данными для добавления результатов"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetAllFinalResults",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"description": "Получить все результаты",
				"tags": [
					"FinalResults"
				],
				"security": [],
				"parameters": []
			}
		},
		"/auth/telegram": {
			"post": {
				"operationId": "AuthenticateTelegramUser",
				"responses": {
					"200": {
						"description": "Объект с данными пользователя и токеном",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"token": {
											"type": "string"
										},
										"user": {}
									},
									"required": [
										"token",
										"user"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Аутентификация через Telegram Web App (TWA)",
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"description": "Объект с initData",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/TelegramAuthRequest",
								"description": "Объект с initData"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "AuthenticateBrowserUser",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"description": "Аутентификация через браузер (GET)",
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": [
					{
						"description": "Хэш запроса",
						"in": "query",
						"name": "hash",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "Telegram ID пользователя",
						"in": "query",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "Имя",
						"in": "query",
						"name": "first_name",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "Фамилия",
						"in": "query",
						"name": "last_name",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "Имя пользователя",
						"in": "query",
						"name": "username",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "URL фотографии",
						"in": "query",
						"name": "photo_url",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "auth_date",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/affirmations": {
			"post": {
				"operationId": "CreateAffirmation",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"description": "Добавить новую аффирмацию",
				"tags": [
					"Affirmations"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateAffirmationRequest"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetAllAffirmations",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"description": "Получить все аффирмации",
				"tags": [
					"Affirmations"
				],
				"security": [],
				"parameters": []
			}
		},
		"/affirmations/random": {
			"get": {
				"operationId": "GetRandomAffirmation",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"description": "Получить случайную аффирмацию",
				"tags": [
					"Affirmations"
				],
				"security": [],
				"parameters": []
			}
		},
		"/affirmations/{id}": {
			"delete": {
				"operationId": "DeleteAffirmation",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"description": "Удалить аффирмацию по ID",
				"tags": [
					"Affirmations"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "https://riqtu.ru/api",
			"description": "Local development"
		},
		{
			"url": "https://hohma-server.ru/api",
			"description": "Prod development"
		}
	]
}